TAP-103: Article header style following the same pattern
