+++ packages/article/styles/const/index.android.js
@@ -1,5 +1,6 @@
-// eslint-disable-next-line import/prefer-default-export
 export const FONT_FAMILY_BODY = "TimesDigitalW04-Regular";
+export const FONT_FAMILY_STAND_FIRST = "TimesModern-Regular";
+
 export const COLOUR_TEXT = "#333333";
 export const COLOUR_HEADLINE = "#000000";
 export const COLOUR_BORDER = "#D0CECE";
